Matching game - 321GO! & Instancing scene
many alterations had to be made to the color application methods because i forgot to do a 2D project. But the effect is the same and no other meaningful code was created. I still use events for color, i just had to include an extra step in acquiring the prefab's material to acces it's color value.